moved theme stuff back into one file, moved ThemeType enum to data layer
