+++ lib/data/adapters/theme_type_adapter.dart
@@ -1,6 +1,6 @@
 import 'package:hive/hive.dart';
 import 'package:silvertimetable/data/hive_type_ids.dart';
-import 'package:silvertimetable/themes.dart';
+import 'package:silvertimetable/data/models/enums.dart';
 
 class ThemeTypeAdapter extends TypeAdapter<ThemeType> {
   @override
