+++ lib/presentation/screens/settings/theme/theme.dart
@@ -1,6 +1,7 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:silvertimetable/data/models/enums.dart';
 import 'package:silvertimetable/logic/settings/settings_cubit.dart';
 import 'package:silvertimetable/presentation/screens/settings/theme/widgets/auto_color_scheme_tile.dart';
 import 'package:silvertimetable/presentation/screens/settings/theme/widgets/custom_color_scheme_tile.dart';
@@ -9,7 +10,6 @@ import 'package:silvertimetable/presentation/screens/settings/theme/widgets/retr
 import 'package:silvertimetable/presentation/screens/settings/widgets/auto_theme_mode_tile.dart';
 import 'package:silvertimetable/presentation/screens/settings/widgets/dark_theme_mode_tile.dart';
 import 'package:silvertimetable/presentation/widgets/category_label.dart';
-import 'package:silvertimetable/themes.dart';
 
 class ThemeScreen extends StatelessWidget {
   const ThemeScreen({super.key});
@@ -41,7 +41,10 @@ class ThemeScreen extends StatelessWidget {
                     CategoryLabel(text: 'colors'.tr()),
                     const AutoColorSchemeTile(),
                     const RetroColorSchemeTile(),
-                    if (state.themeType == ThemeType.custom) const CustomColorSchemeTile() else const SizedBox.shrink(),
+                    if (state.themeType == ThemeType.custom)
+                      const CustomColorSchemeTile()
+                    else
+                      const SizedBox.shrink(),
                   ],
                 ),
               );
