+++ lib/presentation/screens/settings/theme/widgets/auto_color_scheme_tile.dart
@@ -1,8 +1,8 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:silvertimetable/data/models/enums.dart';
 import 'package:silvertimetable/logic/settings/settings_cubit.dart';
-import 'package:silvertimetable/themes.dart';
 
 class AutoColorSchemeTile extends StatelessWidget {
   const AutoColorSchemeTile({
